LED debugging for encoders
Now we keep track of whether the values read from the encoders change
from cycle to cycle.  If none of the encoders have changed their values
we light the 4th LED (the innermost one on the board).  If any have
failed to change their value we light the 3rd LED.

Note that we might see occasional unchanged values by chance.  It's when
the LEDs are continually lit that there's likely to be a problem.