Send the divider to all motors on first position
Previously the code would send a the divider and offset the first time
it set a motor into position control mode, but this meant that only a
single motor would have its divider and offset changed.  Now each motor
has these set the first time it is put in position mode.